RenderActivity: use class loader to detect -main function.
The app caches and attempts to execute from cache (on subsequent invocations)
only if the Clojure program has a -main function.  Instead of scanning the
source code for a (def -main []...) or similar, which is fragile, look up the
names of the classes we collect during compilation. The -main method then
appears as a "clojure.core$_main" (hypens are replaced with underscores when
generating symbols).

Signed-off-by: Iliyan Malchev <malchev@gmail.com>