RenderActivity: set last error on an AndroidRuntime crash
If RenderActivity crashes with an AndroidRuntime crash (due to the
Clojure code), trap the error and set clojureStatus accordingly, so that
it can be fed back as an error code to the next iteration.

With this change, an AndroidRuntime crash both correctly collects the
RenderActivity's logcat output as well as sets the last-error state.

Signed-off-by: Iliyan Malchev <malchev@gmail.com>