DesignSession: introduce initialCode, rework currentCode
ClojureAppDesignActivity was using currentCode as a short-hand for
passing the initial code to a DesignSession.  Introduced a separate
field for this purpose specifically, and updated
ClojureAppDesignActivity.

Also, in DesignSession, this change sets currentCode not from the saved
value in JSON, but from the extracted value in the message history, and
asserts that the two be equal.  A subsequent change will remove
currentCode altogether and extract it on demand from the message
history.

Signed-off-by: Iliyan Malchev <malchev@gmail.com>