ClojureAppDesignActivity: handle sessions with only a system prompt
The logic to restore the selected message was unnecessarily complicated
and wrong in that it excluded the system prompt, causing an assertion
failure with sessions containing only a single message which is a system
prompt.  This shouldn't happen anyway but the code should be able to
handle it if it does.

Signed-off-by: Iliyan Malchev <malchev@gmail.com>